+++ ext/LuxZygoteExt/LuxZygoteExt.jl
@@ -1,0 +1,17 @@
+module LuxZygoteExt
+
+using ADTypes: AutoZygote
+using ChainRulesCore: ChainRulesCore
+using Lux: Lux
+using Setfield: @set!
+using Zygote: Zygote
+
+const CRC = ChainRulesCore
+
+Lux._is_extension_loaded(::Val{:Zygote}) = true
+
+include("training.jl")
+include("nested_ad.jl")
+include("batched_ad.jl")
+
+end
